refactor(v036): split PersonaPicker + agents helpers out of Chat.svelte for size cap (RFC 0048)
The earlier commits pushed two files past the 500-line review-size limit that CI
enforces (file_size.py --strict): Chat.svelte (552) and registry_test.go (526).
Both sat right at the cap on main; the new lobby/exit and sort-test additions
tipped them over.

Split, not trim (the codebase's own remedy — cf. PublishComposer):
- Extract the persona switcher + Exit + lobby into PersonaPicker.svelte, and the
  shared isChattable/agentLabel display helpers into lib/agents.js. Chat.svelte
  552 → 488; rendered DOM is unchanged, so all panel tests pass as-is.
- Move TestListSortedByID into registry_list_test.go (same package, shares the
  test helpers). registry_test.go back to 499.

No behavior change. Web suite 140 passing; registry green; file_size --strict OK.
